moving Origin to ast module
