+++ ast/src/main/kotlin/com/strumenta/kolasu/ast/Origin.kt
@@ -1,0 +1,8 @@
+package com.strumenta.kolasu.ast
+
+interface Origin {
+    var range: Range?
+    val sourceText: String?
+    val source: Source?
+        get() = range?.source
+}
