+++ core/src/main/kotlin/com/strumenta/kolasu/model/Model.kt
@@ -3,6 +3,7 @@ package com.strumenta.kolasu.model
 import com.strumenta.kolasu.ast.Derived
 import com.strumenta.kolasu.ast.Internal
 import com.strumenta.kolasu.ast.Link
+import com.strumenta.kolasu.ast.Origin
 import com.strumenta.kolasu.ast.Range
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty1
