+++ emf/src/main/kotlin/com/strumenta/kolasu/emf/Model.kt
@@ -4,12 +4,12 @@ import com.google.gson.JsonObject
 import com.google.gson.JsonParser
 import com.strumenta.kolasu.antlr.parsing.ParseTreeOrigin
 import com.strumenta.kolasu.ast.Destination
+import com.strumenta.kolasu.ast.Origin
 import com.strumenta.kolasu.ast.Point
 import com.strumenta.kolasu.ast.Range
 import com.strumenta.kolasu.model.NodeDestination
 import com.strumenta.kolasu.model.NodeLike
 import com.strumenta.kolasu.model.NodeOrigin
-import com.strumenta.kolasu.model.Origin
 import com.strumenta.kolasu.model.ReferenceByName
 import com.strumenta.kolasu.model.SimpleOrigin
 import com.strumenta.kolasu.model.TextFileDestination
